keep ts loader config consistent across apps
